♻️ Migrate the abstraction for the external service *ClickUp* into a single module.
