+++ gearmeshing_ai/scheduler/checking_points/clickup/overdue_tasks.py
@@ -9,10 +9,12 @@ from typing import Any
 
 from clickup_mcp.models.dto.task import TaskResp
 
-from gearmeshing_ai.scheduler.checking_points.base import CheckingPointType, ClickUpCheckingPoint
+from gearmeshing_ai.scheduler.checking_points.base import CheckingPointType
 from gearmeshing_ai.scheduler.models.checking_point import CheckResult, CheckResultType
 from gearmeshing_ai.scheduler.models.monitoring import MonitoringData, ClickUpTaskModel
 from gearmeshing_ai.scheduler.models.workflow import AIAction
+
+from .base import ClickUpCheckingPoint
 
 
 class OverdueTaskCheckingPoint(ClickUpCheckingPoint):
