+++ gearmeshing_ai/scheduler/checking_points/clickup/urgent_tasks.py
@@ -7,10 +7,12 @@ attention and triggers appropriate AI workflows for triage and action.
 from datetime import datetime, timedelta
 from typing import Any
 
-from gearmeshing_ai.scheduler.checking_points.base import CheckingPointType, ClickUpCheckingPoint
+from gearmeshing_ai.scheduler.checking_points.base import CheckingPointType
 from gearmeshing_ai.scheduler.models.checking_point import CheckResult, CheckResultType
 from gearmeshing_ai.scheduler.models.monitoring import MonitoringData, ClickUpTaskModel
 from gearmeshing_ai.scheduler.models.workflow import AIAction
+
+from .base import ClickUpCheckingPoint
 
 
 class UrgentTaskCheckingPoint(ClickUpCheckingPoint):
